changed the oldfunctions file
