+++ lib/utils/functions.js
@@ -91,7 +91,6 @@ async function getComponentJson(apiTitle, api, swaggerUrl) {
   return componentJson;
 };
 
-
 module.exports = {
   transliterateObject,
   quote,
